fix(reviewer-sweep): apply critical reviewer findings
Address the open severity:critical / code-review issues from the
2026-05-04 reviewer sweep:

- #706 ui-test: route the WelcomeViewTests hero-absence assertion
  through `WelcomeScreen.expectHeroAbsent(timeout:)` and switch the
  remaining `app.application.buttons["Get started"]` lookups in the
  screen driver to `UITestIdentifiers.Welcome.heroGetStartedButton`
  so the literal isn't leaked from any test-side surface.
- #685 ui: gate `RecentlyAddedView`'s context-menu Delete behind a
  confirmation dialog matching the `TransactionListView` pattern.
- #686 ui: replace the keyboard-inaccessible `.onTapGesture` on the
  Import Rules `RuleRow` with `List(selection:)` plus
  `contextMenu(forSelectionType:primaryAction:)` so Return /
  double-click open the editor.
- #683 ui: give the Earmark savings-progress `ProgressView` an
  explanatory `.accessibilityLabel` and group the surrounding stack
  with `.accessibilityElement(children: .combine)`.
- #682 ui: build a descriptive accessibility label for the Net Worth
  chart from the visible date range and current value, and switch the
  series-toggle buttons from `.plain` to `.borderless` so Space
  activates them on macOS.
- #647 code: replace the `CloudKitBackend` parameter on
  `seedSidebarPreview` (and seven other Features-side preview helpers
  with the same pattern) with `any BackendProvider`, restoring the
  CLAUDE.md "no Backends/ imports in Features" rule.
- #648 code: split `ReportingStore.swift` into
  `CapitalGainsSummary.swift` (with `TaxAdjustmentValues` nested,
  since it's only ever produced by `asTaxAdjustmentValues`) and a
  `ReportingStore.swift` that holds only the store.
- #643 appstore: add `App/PrivacyInfo.xcprivacy` declaring
  `NSPrivacyAccessedAPICategoryUserDefaults` (CA92.1) and
  `NSPrivacyAccessedAPICategoryFileTimestamp` (C617.1) so App Store
  upload validation passes.

Side fix to keep `type_body_length` happy after the chart's
accessibility label addition: extracted the per-series chart-content
builders to `NetWorthGraphCard+Marks.swift`.
